Add generic tests, fix a couple of bugs
	modified:   package.yaml
	renamed:    src/Lex.hs -> src/Clex/Lex.hs
	renamed:    src/Types.hs -> src/Clex/Types.hs
	modified:   stack.yaml
	new file:   test/Clex/LexSpec.hs
	new file:   test/Clex/TypesSpec.hs
	modified:   test/Spec.hs

Added pretty basic tests that generate tokens and then lex to and from
bytestrings.

Fixed a bug where I didn't realise unsigned could be 0 or more
instances, and not a single char.

Fixed a bug where (tokenToBS Not) == "!=", instead of "!".

Fixed a bug where I was parsing Numbers before Floating numbers,
and as a result "0.123" would show up as [Number _, DotFloat _]
instead of [LeadingDot _].